Clean up showcase components
This restructures the showcase components into a new module
(`components.ts`), and splits the implementation into distinct functions
for readability. Moreover the implementation of the anchor picker
showcase component is simplified.
