+++ src/showcase/src/pages/components.astro
@@ -4,7 +4,7 @@ import "$src/styles/main.css";
 
 <div id="pageContent"></div>
 <script>
-  import { componentsPage } from "$showcase/showcase";
+  import { componentsPage } from "$showcase/components";
 
   componentsPage();
 </script>
